Prevent different gut repos from getting mixed by gut-fetching from stale gut-daemons
This adds the tail hash to be part of the gut repo url. Previously, we set up the
remote as git://localhost:port/, which pointed to the gut repo root. However,
if a gut-fetch was called pointing to a gut daemon from a completely different
folder (in my case, I was switching between ~/test and ~/work folders/repos), then
gut-daemon will gladly send data for the "master" branch it knows about, and then
the client will attempt to merge together these disparate histories, which is never
what we want to happen.